# Conflicts:
#   game-project/game/info3/game/modele/MoveableEntity.java

+++ game-project/game/info3/game/modele/MoveableEntity.java
@@ -64,8 +64,8 @@ public abstract class MoveableEntity extends Entity {
 		return m_damage;
 	}
 
-	public void setDamage(float m_damage) {
-		this.m_damage = (int) m_damage;
+	public void setDamage(int m_damage) {
+		this.m_damage = m_damage;
 	}
 
 }
